% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_function}
\alias{hello_function}
\title{Hello world function}
\usage{
hello_function(name)
}
\arguments{
\item{name}{The name of the person you want to say hello to}
}
\value{
a character string of 'Hello {name}'
}
\description{
Hello world function
}
\examples{
hello_function("Jeff")

}
