% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_error.R
\name{squared_error}
\alias{squared_error}
\title{Squared-error function}
\usage{
squared_error(reference, calculated)
}
\arguments{
\item{reference}{vector of the original values}

\item{calculated}{vector of the values we want to compare}
}
\value{
Returns the sum of the squared pair-wise differences between vectors
}
\description{
This is a squared-error function for us to learn the basics of
  package building.
}
\details{
This function is vectorised to give us a helpful example.  Here is
  some text about it to include in the function help.
}
\examples{
reference <- c( 12, 15, 25, 31)
calculated <- c( 10, 17, 26, 30)

squared_error(reference, calculated)
}
