% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_error_plot.R
\name{squared_error_plot}
\alias{squared_error_plot}
\title{Squared error plot}
\usage{
squared_error_plot(reference, calculated, title = "Squared error plot")
}
\arguments{
\item{reference}{The reference values to compare against}

\item{calculated}{The calculated values to be comparared against reference.}

\item{title}{The title for you plot.  Default}
}
\value{
A ggplot object (list)
}
\description{
Squared error plot
}
\examples{

reference <- rnorm(100, 30, 5)
calculated <- reference + rnorm(100, 0, 5)

squared_error_plot(reference, calculated)
}
