% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_error_plot.R
\name{squared_error_plot}
\alias{squared_error_plot}
\title{Squared error plot}
\usage{
squared_error_plot(reference, calcualted)
}
\arguments{
\item{reference}{The reference values to compare against}

\item{calcualted}{The calcualted values to be comparared against reference.}
}
\value{
A ggplot object (list)
}
\description{
Squared error plot
}
\examples{

reference <- c( 12, 15, 25, 31)
calculated <- c( 10, 17, 26, 30)

squared_error_plot(reference, calculated, title)

}
